# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp AI Agent — Multi-Tenant SaaS API
  version: 1.0.0
  description: |
    ## Overview
    Multi-tenant restaurant ordering platform powered by an AI agent over WhatsApp.
    Each **Business** (tenant) connects a WhatsApp phone number, and the AI agent
    handles customer orders, menu inquiries, complaints, and status checks automatically.

    **Two API surfaces:**
    | Surface | Auth | Scope |
    |---------|------|-------|
    | **Tenant API** | `x-api-key` or Bearer JWT | Scoped to one business |
    | **Platform API** | Bearer JWT only (`businessId: null`) | Cross-tenant super-admin |

    ## Webhook Flow
    1. Meta sends a WhatsApp message to `POST /api/webhook`
    2. Server identifies the Business via `phone_number_id`
    3. AI agent processes the message (menu query, order submission, etc.)
    4. Reply is sent back as the response body

    ## Real-time Events (Socket.io)
    - **URL:** `http://localhost:3001`
    - **Rooms:** `business:{businessId}`
    - **Events:**
      | Event | Payload | Description |
      |-------|---------|-------------|
      | `new-message` | `{ customerId, customerPhone, message, reply }` | AI replied to a customer |
      | `new-order` | `{ orderId, referenceId, customerId, items, totalPrice }` | New order placed via WhatsApp |
      | `order-updated` | `{ orderId, status, updatedAt }` | Order status changed |
...

    ## Demo Credentials
    | Role | API Key / Endpoint | Email | Password |
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Business API key (e.g. `demo-api-key-123`)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token from `/api/auth/login`
    PlatformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Platform user JWT from `/api/auth/platform-login`.
        Token contains `businessId: null`.
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        isActive:
          type: boolean
        roles:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              isSystem:
                type: boolean
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    UserInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    UserUpdateInput:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
...
paths:
...
  /users:
    get:
      tags: [Users (IAM)]
      summary: List tenant users
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: role
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Sanitized user list (no password hash)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Users (IAM)]
      summary: Create a tenant user (auto-assigns Staff role)
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email already exists in this business
  /users/{id}:
    get:
      tags: [Users (IAM)]
      summary: Get user with roles and permissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User with roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [Users (IAM)]
      summary: Update user (name, isActive)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateInput'
      responses:
        '200':
          description: User updated
        '403':
          description: Cannot modify own account
    delete:
      tags: [Users (IAM)]
      summary: Soft-delete a user (sets isActive=false)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User soft-deleted
        '403':
          description: Cannot delete own account
  /users/{id}/roles:
    post:
      tags: [Users (IAM)]
      summary: Assign a role to a user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [roleId]
              properties:
                roleId:
                  type: string
      responses:
        '201':
          description: Role assigned
        '409':
          description: Role already assigned
  /users/{id}/roles/{roleId}:
    delete:
      tags: [Users (IAM)]
      summary: Revoke a role from a user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Role revoked

  # ─── ROLES (IAM) ────────────────────────────────────────────────────────────
...
  /platform/users:
    get:
      tags: [Platform - Users]
      summary: List platform admin users
      security:
        - PlatformBearerAuth: []
      responses:
        '200':
          description: Platform admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Platform - Users]
      summary: Create a new platform admin user
      security:
        - PlatformBearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
      responses:
        '201':
          description: Platform user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Platform user with this email already exists
  /platform/users/{id}:
    delete:
      tags: [Platform - Users]
      summary: Deactivate a platform admin (soft-delete)
      security:
        - PlatformBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deactivated
        '404':
          description: Not found
...